organizations/bitbucket-org/requirements.txt
requests>=2.31.0,<3
atlassian-python-api>=4.0.7